// demo-ui/src/app/features/generated/reference/localization/city/models/CityFullView.ts
import { CityForm } from './CityForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class CityFullView {

id: number;
form: CityForm;
canUpdate: boolean;

/* Specific Code Start */
/* Specific Code End */
}